refactoring redis
