+++ handler/twitch.go
@@ -7,12 +7,12 @@ import (
 	"time"
 
 	"github.com/go-chi/chi/v5"
-	"github.com/ruvice/dotabackseaterbackend/repository"
+	"github.com/ruvice/dotabackseaterbackend/repository/redisRepo"
 	"github.com/ruvice/dotabackseaterbackend/wrapper"
 )
 
 type TwitchHandler struct {
-	Redis         *repository.RedisRepo
+	Redis         *redisRepo.RedisRepo
 	TwitchWrapper *wrapper.TwitchWrapper
 }
